requirements.txt
fastapi
psycopg2-binary
pandas
python-dotenv
uvicorn
